+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/cart/WooPosCartScreen.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.woopos.cart
+package com.woocommerce.android.ui.woopos.home.cart
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.layout.Box
@@ -21,8 +21,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.woocommerce.android.ui.woopos.cart.products.ProductSelector
-import com.woocommerce.android.ui.woopos.cart.products.ProductSelectorViewModel
+import com.woocommerce.android.ui.woopos.home.products.ProductSelector
+import com.woocommerce.android.ui.woopos.home.products.ProductSelectorViewModel
 import com.woocommerce.android.ui.woopos.util.WooPosPreview
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
