+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/products/ProductSelector.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.woopos.cart.products
+package com.woocommerce.android.ui.woopos.home.products
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
