+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/products/ProductsDataSource.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.woopos.cart.products
+package com.woocommerce.android.ui.woopos.home.products
 
 import com.woocommerce.android.model.Product
 import kotlinx.coroutines.flow.Flow
