+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/products/ProductsDataSourceImpl.kt
@@ -1,11 +1,12 @@
-package com.woocommerce.android.ui.woopos.cart.products
+package com.woocommerce.android.ui.woopos.home.products
 
 import com.woocommerce.android.model.Product
 import com.woocommerce.android.ui.products.selector.ProductListHandler
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-class ProductsDataSourceImpl @Inject constructor(private val handler: ProductListHandler) : ProductsDataSource {
+class ProductsDataSourceImpl @Inject constructor(private val handler: ProductListHandler) :
+    ProductsDataSource {
     override val products: Flow<List<Product>> = handler.productsFlow
 
     override suspend fun loadProducts() {
