+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/root/navigation/WooPosMainFlowGraph.kt
@@ -3,8 +3,8 @@ package com.woocommerce.android.ui.woopos.root.navigation
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.navigation
-import com.woocommerce.android.ui.woopos.cart.CART_ROUTE
-import com.woocommerce.android.ui.woopos.cart.cartScreen
+import com.woocommerce.android.ui.woopos.home.cart.CART_ROUTE
+import com.woocommerce.android.ui.woopos.home.cart.cartScreen
 import com.woocommerce.android.ui.woopos.checkout.checkoutScreen
 import com.woocommerce.android.ui.woopos.checkout.navigateToCheckoutScreen
